requirements.txt
Django==4.1.3
djangorestframework==3.14.0
django-cors-headers==3.13.0
djangorestframework-simplejwt==5.2.2
django-configurations==2.4
django-phonenumber-field==7.0.0
django-phonenumbers==1.0.1
wheel==0.37.1
Faker==15.1.3
ipython==8.6.0
psycopg2-binary==2.9.5
drf-yasg==1.21.4
factory-boy==3.2.1
django-filter==22.1
celery==5.2.7
redis==4.3.4
tblib==1.7.0
